requirements.txt
requests
discord.py==1.7.2
aiohttp
aiosignal
async_timeout
python-dotenv
wikitextparser